Refs #31814. Add tests for dead time files.
