+++ scripts/Muon/GUI/Common/corrections_tab_widget/dead_time_corrections_presenter.py
@@ -140,7 +140,8 @@ class DeadTimeCorrectionsPresenter:
         try:
             name = load_dead_time_from_filename(filename)
         except Exception:
-            self._corrections_presenter.warning_popup("The file provided has an unexpected format.")
+            self._corrections_presenter.warning_popup("The file provided has an unexpected format. The file should be "
+                                                      "of the same instrument and cycle as the raw data.")
             return None
 
         if name == "":
