+++ scripts/test/Muon/CMakeLists.txt
@@ -9,6 +9,8 @@ set ( TEST_PY_FILES
    corrections_tab_widget/corrections_model_test.py
    corrections_tab_widget/corrections_presenter_test.py
    corrections_tab_widget/corrections_view_test.py
+   corrections_tab_widget/dead_time_corrections_model_test.py
+   corrections_tab_widget/dead_time_corrections_presenter_test.py
    corrections_tab_widget/dead_time_corrections_view_test.py
    elemental_analysis/elemental_analysis_test.py
    elemental_analysis/name_generator_test.py
